Better pdf error handling
